Found issue to be slow image compression. Disabling it for now, and reducing timeouts
#46

+++ rpiDaemon/Jobs/UploadImagesJob.cs
@@ -17,13 +17,17 @@ namespace rpiDaemon.Jobs;
 [DisallowConcurrentExecution]
 public class UploadImagesJob : IJob
 {
-    private const int MaxNumberToUpload = 6;
+    private const int MaxNumberToUpload = 10;
     private readonly IAzureStorageService _azureStorageService;
     private readonly IHostEnvironment _environment;
     private readonly IFileSystemService _fileSystemService;
     private readonly ILogger<UploadImagesJob> _logger;
     private readonly IPictureEditService _pictureEditService;
     private readonly IRepositories _repos;
+
+    // WebP compression takes incredibly long time (~30 sec when running on Raspberry Pi. Seems like it is an issue when running on ARM64 architecture: https://github.com/SixLabors/ImageSharp/issues/2125 
+    // Currently disabling it
+    private readonly bool CompressToWebPFormat = false;
 
     public UploadImagesJob(IAzureStorageService azureStorageService, IFileSystemService fileSystemService,
         IPictureEditService pictureEditService, ILogger<UploadImagesJob> logger, IHostEnvironment environment,
@@ -57,11 +61,12 @@ public class UploadImagesJob : IJob
             return;
         }
 
-        _logger.LogInformation("Found {Count} files for upload. Uploading the first {Max}", imagesForUpload.Count(), MaxNumberToUpload);
+        _logger.LogInformation("Found {Count} files for upload. Uploading the first {Max}", imagesForUpload.Count(),
+            MaxNumberToUpload);
         foreach (var filePath in imagesForUpload.Take(MaxNumberToUpload))
         {
             // Seeing weird behavior with Task being cancelled directly after starting upload. Trying to use another CancellationToken...
-            var uploadTimeout = new CancellationTokenSource(TimeSpan.FromSeconds(90));
+            var uploadTimeout = new CancellationTokenSource(TimeSpan.FromSeconds(60));
             var linkedCToken =
                 CancellationTokenSource.CreateLinkedTokenSource(uploadTimeout.Token, context.CancellationToken);
             var localFileLocation = Path.Combine(imagesFolder, filePath);
@@ -69,10 +74,15 @@ public class UploadImagesJob : IJob
                 $"{filePath}.jpg",
                 $"{localFileLocation}.jpg", linkedCToken.Token);
             _logger.LogInformation($"Uploaded jpg image {localFileLocation}.jpg");
-            var compressedImagePath = await _pictureEditService.CompressJpgToWebPFormat($"{localFileLocation}.jpg", linkedCToken.Token);
-            _logger.LogInformation($"Saved compressed image at {compressedImagePath}");
-            await UploadImageToContainerClient(GlobalConstants.ThumbnailImagesContainerName,
-                $"{filePath}.webp", compressedImagePath, linkedCToken.Token);
+            string compressedImagePath = null;
+            if (CompressToWebPFormat)
+            {
+                compressedImagePath =
+                    await _pictureEditService.CompressJpgToWebPFormat($"{localFileLocation}.jpg", linkedCToken.Token);
+                _logger.LogInformation($"Saved compressed image at {compressedImagePath}");
+                await UploadImageToContainerClient(GlobalConstants.ThumbnailImagesContainerName,
+                    $"{filePath}.webp", compressedImagePath, linkedCToken.Token);
+            }
 
             await _repos.ImageUploads.AddAsync(new ImageUploadModel
             {
@@ -81,14 +91,19 @@ public class UploadImagesJob : IJob
                 FullSizeImageUrl = _azureStorageService.GetBlobUrl(GlobalConstants.ImagesContainerName,
                     $"{filePath}.jpg"),
                 ThumbnailJpgImageUrl = null,
-                ThumbnailWebPImageUrl = _azureStorageService.GetBlobUrl(GlobalConstants.ThumbnailImagesContainerName,
-                    $"{filePath}.webp")
+                ThumbnailWebPImageUrl =
+                    await _azureStorageService.ExistsAsync(GlobalConstants.ThumbnailImagesContainerName,
+                        $"{filePath}.webp", linkedCToken.Token)
+                        ? _azureStorageService.GetBlobUrl(GlobalConstants.ThumbnailImagesContainerName,
+                            $"{filePath}.webp")
+                        : null
             }, context.CancellationToken);
             _logger.LogInformation(
                 "Saved ImageUpload entry {path} to ImageUploads repo. Deleting it and compressed image", filePath);
 
             _fileSystemService.DeleteFile($"{localFileLocation}.jpg");
-            _fileSystemService.DeleteFile(compressedImagePath);
+            if (compressedImagePath != null)
+                _fileSystemService.DeleteFile(compressedImagePath);
         }
 
         foreach (var folderName in folderNamesForUpload)
